+++ test/samples/quoteInQuotes.json
@@ -1,0 +1,1 @@
+{ "name": "Broadcast "Media" }
